fix updating of images
When we create new posts, we add images as File objects to FormData
object and server accepts those files separate from request.body for
both adding and updating.

We fetch the images as plain objects instead of File objects on
YourPosts screen. The content field of a single image is base64 string
rather than a buffer. When we populate an old post form to update it the
form contains the images as plain objects. That's why we convert them to
File objects before sending an update request.