change solution for unique key issues
We don't allow adding duplicate links in both frontend and backend.

Using the uniqueId function as a key had issues with the FileUploader
component. Multiple clicks were required to delete the dropped image
with the button.

Mongo stores _ids for each image in the images array. Thus, we use these
_id values when displaying images retrieved from backend.