refactor the central error handling
We deleted the custom error class that extends the regular Error class.
Instead, we used the regular Error class and append custom fields to it.
This also eliminated the previous stack trace issue.

Errors that occur in lower layers such as services or repositories are
thrown to the controller layer without a try/catch block. The thrown
error object is modified in the catch block of the controller if it is
necessary and moved to the handleError middleware. Finally, the error is
logged here and an error response is returned to the client.

We use the description field to return client friendly error messages.
The default message field is useful for identifying errors that not
explicitly thrown.