refactor image uploading and saving processes
We changed the multer storage type to disk storage.

There is a public folder on the server and we save the incoming images
to the public/<IMAGE_UPLOAD_SUBPATH> folder. Now, the paths of the
images on the server are stored in the database.

When a user updates a post, the req.body object contains the paths of
old images while the req.files object contains the file objects of the
new images.

There is no authentication for the public folder. We can use longer
hashes for image names temporarily.

When a user deletes a post, the images of that post are not deleted from
the public folder.

There must be public and public/<IMAGE_UPLOAD_SUBPATH> folders on the
path declared by <PUBLIC_UPLOAD_PATH>.